Revert "CLTool - Improve Linux and Windows cmake process (#386)"
This reverts commit 47aff36b9226dbb38b541289b5d844109cbf83c0.

+++ src/ISCommDataBuffer.cpp
@@ -141,9 +141,9 @@ int cComDataBuffer::ReadData(int pHandle, uint32_t dataId, vector<uint8_t>& data
 #else
     fstat(_fileno(file), &buf);
 #endif
-    size_t size = buf.st_size;
-    size_t pos = ftell(file);
-    size_t afterWriteOffset = size - pos;
+    long size = buf.st_size;
+    long pos = ftell(file);
+    long afterWriteOffset = size - pos;
     size_t read = 0;
 
     // read everything after write offset
